+++ src/com/openscg/pgstudio/client/models/FunctionInfo.java
@@ -14,13 +14,13 @@ public class FunctionInfo implements ModelInfo, Comparable<FunctionInfo> {
       }
     };
 
-    private final int id;
+    private final long id;
     private final int schema;
     private final String name;
     
     private String identity;
     
-    public FunctionInfo(int schema, int id, String name) {
+    public FunctionInfo(int schema, long id, String name) {
     	this.schema = schema;
         this.id = id;
         this.name = name;
@@ -39,13 +39,13 @@ public class FunctionInfo implements ModelInfo, Comparable<FunctionInfo> {
       return false;
     }
 
-    public int getId() {
+    public long getId() {
       return this.id;
     }
 
     @Override
     public int hashCode() {
-      return id;
+      return (int) id;
     }
 
 	public String getName() {
