+++ src/com/openscg/pgstudio/client/providers/PolicyListDataProvider.java
@@ -23,13 +23,13 @@ import com.openscg.pgstudio.client.models.PolicyInfo;
 public class PolicyListDataProvider extends AsyncDataProvider<PolicyInfo> implements ItemListProvider {
 	private List<PolicyInfo> list = new ArrayList<PolicyInfo>();
 
-	private int item = -1;
+	private long item = -1;
 	private int schema = -1;
 	private ITEM_TYPE type;
 
 	private final PgStudioServiceAsync studioService = GWT.create(PgStudioService.class);
 
-	public void setItem(int schema, int item, ITEM_TYPE type) {
+	public void setItem(int schema, long item, ITEM_TYPE type) {
 		this.item = item;
 		this.schema = schema;
 		this.type = type;
